Changes to build and README
Refactored how Solidity contracts are built.

Reimplemented the shell script to compile Solidity contracts to be
more robust and also user-friendly.

Added CMake function that can conveniently be used to create a target
which will take care of all the custom commands needed to build a
specified Solidity source file into bytecode and embed that bytecode
into a generated C++ header file.

Made changes to the README file.
Improvements to the description for the dependencies.
Added brief description for the erc20 contracts.